+++ rest_email_auth/signals.py
@@ -1,0 +1,11 @@
+"""
+Custom signals used by django-rest-email-auth.
+
+These signals allow other applications to easily add custom behavior to
+processes from this module.
+"""
+
+from django import dispatch
+
+
+user_registered = dispatch.Signal(providing_args=['user'])